Fix for technically incorrect statement in the position: absolute vs fixed question.
I didn't want to have to resolve merge conflicts again so I'm doing it as part of this pull request.

It previously said that `position: absolute;` was relative to the **parent element**. This is not true if the parent element is set to `position: static;` (the default).

I've reworded the bracketed text to not confuse juniors. I also fixed some minor grammer errors in that same block of text.